Fixed a page fault bug in kenrel do_unmap_page
Currently, each process's stack is set up asynchronously, when loading it
into memory. This requires unmapping the currently loaded stack first, but
this caused a page fault if it was already unmapped for whatever reason (probably
if the previous process just exited or if the kernel idle process was executing).

Fix this by checking to make sure everything is mapped in before unmapping an
address.