Sh: Allow fg for pipelines
Each job now has a number of processes, and a number of processes
for which it consumed a sleep status. This way, a pipeline can be
put to sleep again correctly, even though most of these sleep messages
are left over from killpg(pgid, SIGSTOP).

Also clean up once a job finishes completely, and print the job status
more consistently.