Kernel: change from fs_clone to fs_dup
Instead of cloning the file data, files are ref counted.

Also, I am not sure if this is the intended behavior across forks,
but it would be difficult to implement it this way, because the
files would need to keep track of their cloned fds, or the model a
process uses would have to change (it could be a list of files
that match more than 1 fd).