sh: support ;, &&, || lists in foreground only mode
Since pipelines are also changed to be single length lists,
no commands can be run in the background.

Eventually, running commands in the background will be accomplished
by starting the list as a shell sub process (if there is more than
one element in the list), allowing this sub process to manage
the list in the background.

Parsing the list is simply a matter of spliting the line
on the characters of interest.