userland: allow compilation of utilites to the native compiler
By creating a new make target `native`, each utility can be
built seperately on the host operating system for more
convenient testing.

This required adding a define `USERLAND_NATIVE` for sh because
it relies on a custom library feature of wordexp, so this is
disabled when compiling on this mode. Eventually, the shell
should not rely on this. Although it makes sense for now.