userland: add head program
This program prints the first n lines from a file,
defaulting to 10 but accepting the parameter from
the -n option.

Internally uses getline to get each line, but since this may
not add a trailing \n character we have to check for eof before
so we add the trailing \n ourselves.