libc: tmpfile and tmpnam
tmpnam uses a shifts the name buffer over to get a new name.
Potentially, it should do something sophisticated. Also, it
only changes the starting point based off the pid, but it only
generates the same thing, just offset. Also, the tmpfs doesn't
exist so the files are not deleted and instead stored on disk