kernel: ping google's ip successfully
NOTE: to get this to work, you have to give your linux group
ping permission by modifying /etc/sysctl.conf

The E1000 now pings to the outside world once it figures
out the routers MAC address.

The kernel network stack now supports ICMP and IPV4 packets
to make this work, and can compute the IPV4 checksum
correctly.

Currently, our IP address is hardcoded to be the same one
passed to qemu, but there should be a way to figure is out
ourselves, a long with the router's IP address.

Also, the code needs to be decoupled from the e1000 driver.
Probably, there is going to be functions to handle each
type of packet that will read the info and then pass
it along to something higher on the network stack.