kernel: socket sys call
Right now, sockets subvert the concept of an inode, but maintain the
idea of a file, and so should be able to integrate well into the
current system. Sockets don't store any buffers or address right
now since nothing else is implemented, but the socket object can
be created and destroyed.