kernel: support for simple tcp interactions
Now, when tcp sockets close, they will send a fin if necessary. Also,
one can recieve from a tcp socket, which will prompt a sending of an
ack. However, sending (when non blocking) does not wait for an ack from
the destinate. Basically, we never resend data, so if something goes
wrong in the connect, we will probably just get spammed by the
destination with acks to tell us to fix our sequence number, but
this actually works in the best case.