userland: start program that executes on startup
Before, we just ran /bin/sh. Now, we run a dedicated c program
that just all the startup things, including giving ownership of
the tty to the shell (before this was hardcoded). Also, the
dns lookup server is created, so now ping URL will work before
starting the server in the shell.

Also, there is a new ioctl, TISATTY, which will return 0 if
the thing is a tty (otherwise -1 w/ errno set to ENOTTY). This
is useful since usually ioctl's will generate SIGTTOU, which
included the isatty call (which was implemented with tcgetattr
and checking errno for ENOTTY). Now, the isatty() calls that
determine whether or not stdin/stdout are _IOLBF or _IOFBF don't
cause the process the go to sleep. The TISATTY ioctl is handled
separately, so this problem is avoided.