kernel: fix ptmx slave_write when writing excessively
As a stress test, I was running grep -r i /usr/include in
the shell. For some reason, the code randomly page faulted in
the kernel, for what seemed like no reason. Eventually, I figured
out it page faulted in memcpy, so I added a debug version that
prints who called it (because making a stack trace is too much
work). Of course, it was being called by realloc (so realloc needed
messages too). Even after limiting the buffer size to 4096, since
before it grew to be like 20K, it still page fauled at realloc.

SOLUTION: remove the realloc code...

I don't understand but there must be something wrong with realloc
(or malloc) that breaks on lightly large buffers.