kernel: add process table
Now, processes are still referenced by individual tasks, but are
ref counted such that there could potentially be multiple tasks.

NOTE: the actual memory of a task is freed when it should be
part of the process instead.

Also, fix use after free bugs in the kernel by destroying the memory
after it is freed in malloc (this exposes the bugs). The most glarying
one was in execve, which freed the global variable current_task, and
then called sys_sched_run_next, which effectively saves the current
task state into that variable. This corrupted memory.

Also, hash_map was suspiciously not freeing the entries it had
created in the method hash_del, this is now fixed.