kernel: store cr3 in a separate struct from the current task
This makes it possible for tasks to share the same cr3, which
this obviously must do to share memory. Now, the kernel stack
of each task is unmapped separately from the rest of the address
space, which happens in proc_kill_arch_process. This means that
tasks could now fully exist in theory (obviously some sys calls
need to be added to create them), but the data is now correctly
dealt with (although fields on process may need to be locked
when doing sys calls as well).