pthreads: create and exit
pthread_create and pthread_exit mostly work now. Obviously, attributes
are completely ignored right now.

Also, needing to make pthreads a separate library is very annoying
because it means libc can't use the locking primitives of pthreads.
For now, I just copied and pasted the code over so that malloc will
be thread safe, but that's a terrible solution