kernel: load tls master copy into process memory
This region is loaded before the executable, so that it cannot
in any way interfere with the heap. The location, size, and
alignment are made available to userspace through the system
call: get_initial_process_info, which also makes the initial
task's stack available, since that is nice to have. Also, the
set_thread_self_pointer system call is added so that when libc
eventually calls the former system call and allocates its tls,
it can tell the kernel what to put in the msr FSBASE. Note that
this should be added as a parameter for create_task, but currently
isn't because it is extra work.