kernel: don't save user task state on task object
Instead, a pointer to this information can be saved, since this
information is only valid when a task is in the kernel, and it
is pushed onto the stack by the system call handler. This is the
start of the process of refactoring the system call code so that
the kernel will only be interrupted by a signal in predefined places
(when block, sys entry, and sys return). This will fix the memory
leaks that currently happen when the kernel is interrupted by a
signal.