kernel: make system calls non interruptible by signals
This finally fixes memory leaks if the kernel is interruped in
system calls like open, which malloc a path string that is
concatenated with the current working directory. This should be
changed to not malloc at all, but that's beside the point.

Now, the scheduling has no RUNNING state, only INTERRUPTIBLE
and UNINTERRUPTIBLE. The RUNNING state is completely useless
since the global variable current is always set to the correct
task anyway, so it just made bookeeping more difficult.

When a system call blocks, it will call either kernel_yield or
__kernel_yield, depending on its needs. kernel_yield will make
the task INTERRUPTIBLE and restore the UNINTERRUPTIBLE status
upon the next scheduling, while __kernel_yield doesn't change
the sched_state field at all, which is used for blocking in
sleep and alarm, where the task is put in the WAITING. All tasks
should probably block like this, so this will be changed and the
unclear naming will be fixed.