pthreads: posix compliant thread specific data
The pthread specific data api is now supported.
There is a bitmap with PTHREAD_KEYS_MAX bits that
keeps track of what keys have been allocated, and
a static array of destructors that will be set with
pthread_key_create.

The implementation is extremely wasteful. Since it always
allocates 1020 bytes per thread to use as an array of
thread specific data. This is however the easiest implementation
possible, and anything more complicated would probably have
a lot of races.