pthreads: cancellation state with signals
Threads can now be cancelled, except there is no cancellation
handler yet. They are delivered through the __PTHREAD_CANCEL_SIGNAL
signal. This should only be unlocked when calling a function marked
as a cancellation point, and when PTHREAD_CANCEL_ASYNCHONOUS is set
as well as PTHREAD_CANCEL_ENABLE. The cancel handle will probably
call pthread_exit, but it needs to make sure to only cancel a thread
that was interruped before a system call finished, or else it could
leak process resources.