kernel: pass other arguments to signal handlers
When the kernel runs a signal handler, it will now
be passed valid pointers to a siginfo_t and ucontext_t.
The siginfo_t isn't initialized to anything yet, but the
ucontext_t has valid pointers to all the registers.

This makes implementing things like real time signals
easier, since this mechanism is already in place.