kernel: fix wrong stack when handling signals
This code dispatched outgoing signals after it had marked the
in_kernel flag to false. This meant that signal handling code
thought the task was already in userland, and used the stack
saved on the task instead of the stack saved on the kernel's
stack, causing the signal handler to execute on the kernel's
stack, page faulting the process.