kernel: create new mechanism for blocking
Since context switching is expensive, busy waiting on conditions
is especially terrible, since more instructions will be wasted context
switching than checking something like whether it is time to wake up
from a sleep.

Because of this, when tasks block, they will provide a simple callback
function called should_unblock, which will be called by the scheduler
every time it tries to schedule that task. If it returns true, the
scheduler will unblock the task. Information needed to make this
decision is stored on the task itself, in the block_info struct/union.

This info cannot be stored on the kernel stack since the entire point
is to avoid context switching. By making the type a union, hopefully
is won't be too space inefficent.

This mechanism is currently only used by sleep but it should be
extended to all other instances where kernel_yield is used to
poll a specific condition and stop running if it is not met.