kernel: select+pselect
Using the previously added readable/writable/exceptional flags for
inodes and sockets, these functions will check for this status on
every file descriptor selected. This uses the newly implemented
blocking mechanism to do the waiting.

Since the file descriptors to watch need to be remembered, the system
call the fd_set's onto its own stack, and works on those. Specifically,
it will unset the particular bit if the condition becomes true, and
xors this with the user-provided fd_set at the end of the function.

This means that the blocker needs access to the kernel's stack when
checking if it should unblock. This is solved on x86_64 by just
using the physical address map to bypass the need to page anything,
but on other architectures the kernel's stack probably needs to
be mapped in to do the check.