kernel: fix sending synchronous signals to our own task
The code used to assume that it could block when waiting for
signal delivery. This is not true, since the action to be
tacken for a signal could be the default one, which doesn't
change the scheduling state of the task. Therefore, the signal
should be executed synchrounously, which is it now.