pthread: framework for robust mutexes
The mechanisms for the implementation are now in place, but it doesn't work at all.
My first attempt was to query the kernel after a lock fails, in order to see if the
process holding the lock wakes up. The issue with this is that a thread will sleep
forever, since the thread that dies never wakes it up, and thus it can't check to
see if the thread that died still exists