libc: fix apparent bug in wordexp we_unescape
This caused inputing quote characters to sh to result in a crash.
Interestingly, this does not happen when wordexp.c is compiled natively,
only when running the actual OS. This most likely means there is some bug
in malloc that caused it, not the code itself that needed was 'fixed'.