wordexp: do arithmetic expansion correctly
Before, only the current operator on the op stack was checked to see
if it had high enough precendence, but now this occures in a loop. This
is important in nested expressions, like 1*2-3/4+5*6, which before would
evaluate this in a right associative manner (since the stack is popped
at the end, and precedence was not respected after an expression was
evaluted). So this would result in 2-0+30, but since this would be
reduced from right to left instead of left to right, the 30 would end
up negative.