kernel: give vm_regions vm_objects
For vm_regions to be shareable and map partial regions,
the code needed become more sophisticated.

A vm_object has a pointer to a vm_object, which provides
a method to map the region into memory. Eventually, mapped
inodes will be stored in a vm_object, which then can be shared
with MAP_SHARED and mapped partially, although this is currently
only used for direct physical mappings.

All instances of vm_objects should be created with vm_create_object,
and when another region maps to the same vm_object, bump_vm_object
should be called. drop_vm_object should be called when a region is
destroyed.

There are currently only two operations that a vm_object does:
  map: which maps a region based on the vm_object and
       region->vm_object_offset.
  kill: called when the object's reference count becomes 0.

Inodes now have a vm_object field, which for now is lazily
initialized once someone tries to mmap it (For mmapable inodes).
This may need to change: for example, when reading from disk
ext2 currently caches blocks in a separate mechanism. It would
make sense to read these directly in the vm_object, since that
would improve efficency when mmap that inode later, and also make
keeping the vm_object up to date with disk much easier.