kernel: make tnode's ref counted
Now that tnode's live outside the file system (they are referenced by
process::cwd), they need to be ref counted so that they aren't
pointlessly copied. This allows code that interacts with the file system
to harmlessly use this field. Before, code that calls fs_symlink would
make the newly created inode's tparent the cwd in certain cases. This
later meant that interacting with this created inode would cause issues,
since the tnode it referenced was freed.