libc: have syscall() verify the number of arguments
Add a _Static_assert so that libc code has to call system calls
with the correct number of arguments. For other userspace code,
we disable this check, since the check is only possible when the
first argument is exactly SC_*. Userspace code should be able to
pass arbitrary variable names to this macro, so this form of check
would break completely.