parser/libliim: make usable
A bug in create_item_sets (taking a reference to a deleted object),
caused everything to be completely broken. However, when using std::
shared_ptr, the undefined behavior miracurously worked, despite
leaking excessive memory. the libliim implementation is not able
to perform this feat, and so parser was always broken.

Fixing this bug allows the libliim pointers to work, and ends
all of the memory leaks in the program!