wordexp: handle parentheses in arithmetic expansion non-recursively
Before, code evaluting terminals would use recursion to evalutate any
expressions inside parentheses, but this is prone to stack overflows.
we_arithmetic_expand creates operator and value stacks, on the stack,
and calling the function recursively wastes tons of memory. Instead,
consider parenthesis an operator with the lowest possible precedance,
and evaluate until one is encountered, when a right parenthesis is
encountered. This algorithm is much more efficent, and makes switching
to a dynamic stack later much less costly.