libc: fix regex group capture with alteration
Regexes like '(ab)c|ab' would match group 1 when given the test
'ab', since the capture from the first group was never cleared,
even though that branch failed to match. This is solved by having
the compiler know which groups are valid for a match, and giving
this information to the RegexState object, which will destroy impossible
to achieve matches when it is transitioned to. This also means that
invalid backreferences like '(ab)c|\\1' can know yield a compilation
error.