wordexp: add do_command_subst callback to word_special_t
The naive implementation simply used popen. This is bad since
popen uses static memory, so it can't be called by libc anyway.
But more importantly, aliases and function names are not preserved
across the execl. This is sovled by forking and then calling the
callback (provided for sh), which can then use the callback to run
the command like normal.