kernel: posix timer_* functions (in progress)
The data model for the timers is in place, all that is left is
to actually tick the timers (right now they don't do anything),
and also handle the actual process of responding to the timer
activating.

Also, add librt for these functions, since they need access to
pthreads they can't be in libc, but there aren't pthread functions
either.