kernel: store timers in per process linked list
This way the timers don't need to be explicitly
timer_delete'd. This uses a fancy trick to allow
insque/remque to work for multiple lists stored
in one struct. Rather than pass in the object,
we pass in &timer->proc_next, so that these functions
operate on the correct list.