kernel: fix socket data race
net_send_to_socket puts the socket_data object into the socket's
queue of messages. Once this object is in the socket's queue, it
no longer owns the socket_data, which will be freed by the socket
once its done with it. This caused a use-after-free bug where the
function actions the socket_data after unlocking the socket's queue.