kernel: allow arbitrary block sizes in ext2 fs
The prior mandated block size of 1024 was extremely inefficent, since
it meant that each 1024 bytes of memory needed to be read at once. This
fix allows for the default 4096 block size, which greatly increases disk
performance. For example, loading gcc's cc1 binary from disk now only
takes around 10 seconds, instead of 30 seconds (which is still
terrible) (oh and it breaks the system when executed).

The main reason this wasn't supported was just that the block group
descriptor table's block offset is variable.

Also, the block caching mechanism prevented this from working easily,
since it caches the super block and block groups, which shouldn't be
cached in the caching system because they need to always be present.