kernel: make hash_map resizable
In the ext2fs, we created a 25 bucket hash_map to cache 1000 blocks.
This was literally so bad that cache lookup ended up being slower than
just reading from disk, which defeats the purpose of a cache. Now
hash_map will automatically resize, so this problem is avoided. And now
that hash_maps' have variable bucket size, we make the cache start with
1000 buckets so that resizing won't happen anyway.