libc: call global constructors/destructors correctly
Calling _init and _fini is supposed to do this, but it actually doesn't.
You have to iterate over __init_array_start to __init_array_end in order
to actually call the global constructors. To do this, we ditch the old
link script hack used when compiling and instead use the base scripts.
This gives access to these symbols, which are definitely necessary.

Now exit() properly calls _fini and global constructors (before this
only happened on return from main). Also, disable -g when compiling
since it only slows the system down (it is extremely disk I/O bound).

This fixed an issue in gcc that made it seg fault when reading from a
statically allocated data structure, but gcc still doesn't actually
work (now it faults in lrealpath).