kernel: fix page fault when doing nothing
The kernel idle task should not just immediately call sched_run_next.
If there is nothing to do (which there probably isn't), sched_run_next
will simply jump back to the idle task, except this time with an
increased rsp. This meant that if you did nothing long enough, the
kernel would page fault on the stack. This is fixed by making the idle
loop actually idle, and this behavior can probably be optimized to check
some global flag to see if it should call sched_run_next (this flag
prevents infinite recursion).