window_server_test: move into userland
For whatever reason, this program doesn't work when put on the initrd.
This is because the initrd is completely and utterly broken. Luckily,
it is completely useless now that the ext2 file system is supported, so
there is no reason to fix it. This does mean that nothing should ever be
added to the initrd again, since it most likely will randomly break for
no apparent reason.