procfs: create read mechanism
Each inode in the procfs has a function pointer to a function that
takes a process and returns a procfs_buffer (char* + size_t) as
information. Right now this is immediately freed, but it could be
cached instead. The buffer is page aligned so that it can be mmaped
aswell.