sh: fix suggestions
When a suggestion is given, it is always relative to some starting
point in the buffer. For instance, a suggestion when given `cat comm`
might be `command.cpp`, starting at index 4. This causes issues, since
this is supposed to overwrite the buffer starting at index 4. When not
at the end of the line, this can mess everything following `comm` up.

The fix is to trim the suggestion to be as minimal as possible. That is,
return "and.cpp" at index 8 instead. This conversion is done after
suggestions are generated for now, but really it would make more sense
to always give suggestions that can be used easily.