kernel: always refresh inodes if needed; do so in procfs
procfs inode's have dynamic content, so their size can change without
there being any writes. The vfs overrides the lookup function to refresh
inode information, but only did so when an inode's size was 0 (since
only ext2 needed to be refreshed, and only once), but now it will always
call this function.

procfs has to generate the content to know's the inode's size, and since
we don't cache things yet, this is very inefficent. The code is now
designed so that this can be changed in the procfs_get_data and
procfs_cleanup_data functions.

Also, store a bitmask in the low bits of the function pointer, so that
the code can differentitate between files in a pid directory, and files
globally available.