kernel: make procfs lazily loaded
Now directory also have a callback in private_data that causes their
contents to be loaded. This means that reading a directory will
unconditionally call lookup, in order to make sure the contents are up
to date. Note that /proc/pid isn't dynamically loaded, the contents of
these directories are loaded on demand. This is actually a fairly
reasonable trade off between wasting space vs computational complexity.