ls: fix symlink resolution
The code assumed that the best way to check if a symlink was valid was
to call access on the result of readlink. This is completely wrong in
the case of relative symlinks (i.e. /proc/self). Luckily, this can
easily be fixed by just calling access on the path that was lstat'ed
directly, which happens to make more sense than the alternative.