kernel: add /proc/pid/vm to look at virtual memory
This change is actually amazing, since one can now actually look at what
the vm_allocator is doing. This patch includes a function
vm_type_to_string which should be used in all memory dumps, and exposes
that stack guard pages are mysteriously mapped with the ability to be
executed.