kernel: make procfs only file size if possible
When stat-ing a procfs inode, the actual contents aren't really needed.
This means the code doesn't need to allocate a buffer. Using snprintf
with n=0 will return the size of the buffer needed, so this convenient
mechanism is used. Interestingly, libc handled this case by accident,
by always pretending writes work even if they go past the end of the
buffer.