kernel: make O_CLOEXEC actually work
The kernel throught it was a good idea to discard this flag every time a
process forks. The whole point of this flag is so that file descriptors
won't be cloned accross an exec (which is almost always preceded by a
fork).

Also open /dev/fb0 with O_CLOEXEC in xterm, so that it doesn't leak the
file descriptor to every single program spawned by the shell (thanks to
/proc/pid/fd for pointing this out).