kernel: rework vfs (and break EVERYTHING)
In order to support hard link's, inodes can't store their parent
directly. The code used to assume that an inode has exactly one parent,
but this is not true in practice. Tons of code was searching for the
name of an inode, which it should not be, since its not supposed to have
one.

The fix is to store the parent chain on tnodes (tree-nodes) instead
of inodes. Also, tnode's can't be used as the one source of truth with
regards to an inode's path.

This current code is terrible and needs to be fixed, but it is 1 A.M.
Basically, the entire cached_dirent code needs to be re done (and I just
wrote it), since all of the current problems can be solved by making the
fs_lookup_in_cache function return a tnode instead of an inode. This
also makes more sense, is more performatant, and still allows hard links
because there will simply be two tnode's pointing to one inode.

The only good thing about this change is that it gets rid of struct
tnode_list, which was objectively terrible. Making path name lookup use
a hash_map instead of a linked list is a legitimitally good change, but
since the current code leaks memory for ever tnode created and causes
ls -l to print 0 directories always, there is definitely room for
improvement.

Also, the kernel hash_map needs to be changed to use closed addressing,
since constant linked-list node allocation makes using hash_map's
painful.

In terms of code reuse, this patch significantly reduces it, by
generalizing directory lookup in a way that all file systems can apply
by simpling calling the appropriate method.