xterm: render the cursor for window server mode
Since we don't store any state beside the raw pixel information, this is
actually quite difficult. We try to guess the foreground color, but this
doesn't work for blank cells. The trick to not destroy state is to save
what the square looked like before we render the cursor, then render the
cursor, and finally restore afterward. This approach does actually work
though.