kernel: fix tmpfs mmap
The system was crashing for some time now, and I have finally found the
root cause. The tmpfs mmap function created a physical inode object that
maps directly to the contents of the in-memory inode. This seems very
efficent since it eliminates the need to copy the inode's contents into
separate pages. This is why tmpfs's buffers are allocated with
aligned_alloc(PAGE_SIZE), since that is necessary for a safe mapping to
exist.

The bug is that a physically mapped inode assumes all memory is
physically continuous. The tmpfs uses aligned_alloc, which only guartees
the contiguity of the virtual addresses allocated. This caused the code
the map arbirtray kernel pages into userland, which preceded to use them
as a graphics buffer, zeroing the memory, and thus causing the kernel's
own page tables to become null.

This is fixed by using an inode_vm_object instead, which does not assume
a physically continuous mapping.