kernel: fix shebang handling (and add test script.sh)
Shebang's are very annoying and obtuse, so their current implementation
is atrocious. This patch adding some strdup's for the recursive path
argument, and will strdup the program name if needed. This makes freeing
the eventual program name very annoying as well, but this will be fixed
when process->name reflects argv[0] rather than exe->name. This is the
first example I've realized where exe->name != argv[0], since the
shebang ruins all that.