xterm: ending a shell session will load another open one
When pressing ^D to exit out of a shell, you presumably want to go to
the next open tab. If there is none, we default to the restarting the
first tty, but it would be better to close the application if running in
graphical mode where there can be other windows.