kernel: fix alarm(2)
So before I had no idea what this function does, so the system call just
slept for the argument's duration and then fired SIGALRM. But its
supposed to register an alarm that will generate a signal
asynchronously.

Luckily, there is already a userland timer subsystem built for sending
signals after a timer elapses, so this was actually trivial to
implement, and now it works correctly.