kernel: make ext2fs support larger inode size's
This patch makes the ext2fs driver aware of the fact that the size of
an ext2 inode is not always 128 bytes. In fact, mke2fs makes the inode
size 256 when the disk image is big enough. This caused the file system
code to always crash when the disk image was too big, which prevented me
from running programs like gcc and binutils with debug symbols available.
Now, I should be able to see stack traces for these ports and figure out
why gcc is broken.

This method is unfortunately very hackish, because the variable nature
of this structure means that c's type system doesn't support it. This
makes pointer arithmetic completely broken, but the code used to rely upon
it. Its fixed by doing casts to char* and doing the raw byte offset, but
this probably should be fixed with a macro and maybe some way to make gcc
warn about using pointer arithmetic with this type.