kernel: report mouse movement in PS/2 mouse driver
The mouse packet you get when reading the mouse device now contains
the dx and dy for the mouse's movement.

This is given directly by the hardware, so we just have to read the
fields from the mouse packet. It is moderately annoying because the
numbers are 9 bits wide, but you read the mouse byte by byte. Also,
the mouse can overflow, in which case we have to ignore the mouse
movement for that event.