kernel: make vm_objects allocate lazily
vm_region's that were backed by an vm_object used to be mapped in as
soon as they were created. This is undesirable for many reasons,
including wasting memory and having to wait for an entire inode to be
read before being able to map it.

The solution is to make vm_object's backing physical pages be loaded on
demand, triggered when a page fault occurs. To do this, the handle_fault
method is added in the vm_object's operations struct.

Also, page fault's used to be handled on a separate stack, but this
broke everything (very mysterious), so they are back to being on the
same stack that caused the fault.

Inode also needed a new function, read, to be able to fill the
inode_vm_object's on demand. This exposes how terrible the vfs code is,
since this function already exists on files, and was already changed to
not rely on the file in any meaningful way (the file's offset is ignored
in order to support the pread and pwrite system calls).