kernel: fix allocation of stack regions
The code called find_user_vm_region_in_range incorrectly, which meant
that it always mapped regions on top of each other. This resulted in
a segfault when creating new threads, since it allocated the stack
directly on top of the guard page, which is mapped PROT_NONE and so
cannot be used as a stack.