kernel: implement wait_queues's
This makes it now possible to wait on arbirtrary events much more easily
than before, with the blocking system. When a task blocks, it needs to
check before every reschedule to see if it should unblock. With wait
queues, this is completely unecessary, and so are more convenient.

Also, these will be used to implement kernel level mutexes, which are
much more efficent than the spinlocks currently used throughout the
kernel.

To test this mechanism, the ata driver was updated to use a wait queue
to notify reading tasks about the occurence of the irq (indicating the
ata drive recieved data).