kernel: allocate stacks completely lazily
The code used to rely on the stack being mapped into memory when passing
the program arguments on the stack. This fix causes a page fault handler
to execute when trying to map in these args, and then the stack will be
mapped in. This means that now the combined size of envp and argv can
now exceed 1 page without crashing the kernel, and there is no memory
leaks associated with the allocation of the stack.