kernel: remove the inode store
This code was made completely useless by the addition of a file's
ownership of a tnode, which directly has a pointer to the inode. This
change has large benefits, as the inode store was often the source of
bugs as it was unclear when to actually put an inode into it to begin
with, and is now made obsolete.