kernel: fix physical page leak in unmap_page
The free_phys argument is used specifically because most physical pages
are manually managed by vm_object's, so can be shared and cannot be
unmapped when its time to clean up the virtual address space. However,
the meta pages directories are not managed in this way, and so must be
freed when pages are unmapped.