kernel: switch execve to load the new image with an inode vm object
The execve function is now split into two distinct parts. The first is
resolving the path and checking the executable's validity and if it
starts with #!. Once the process image we want to load is finalized, we
allocate a new task and process to take over the current one.

There is an atomic switch between the two (interrupts are disabled), and
then the new process is running inside of a system call handler. At that
point, the process image is mapped into the address space and the elf
code finishes the cleanup.

This exposes the fact that a separate process allocation should not be
done, instead the task's state should be transferred over directly
without the need for a complex switching mechanism and pointless
allocations.