kernel: make the heap backed by an vm_object
Since malloc() uses sbrk to allocate its data, the addition of
vm_object::extend was needed to allow a vm_object to be resized. This
now means that _all_ of process memory is backed by a vm_object, so this
should reduce the amount of special case handling in the code.