kernel: rewrite the cloning of address space used by fork()
Now that vm_region's are backed by actual objects, we can easily and
precisely clone them on a fork(). This patch also fixes
soft_remove_paging_structure(), which decided to always deallocate
memory, which it should never do now that the vm_objects keep track of
physical memory usage.

However, this code should be temporary, as implementing COW would be
ideal. To do so, vm_objects would need to store something like a struct
page, which contains a physical address and a reference count, although
it is unclear to me how this would interact with vm_objects.