kernel: fix unmapping of each task's kernel stack
The handling of the kernel stack is very special, since _every_ task
shares the same virtual memory for its kernel stack (this is probably
very vunerable to attacks). Because of this, every time a task is loaded
the kernel stack has to be unmapped and then reloaded. This means that
these unmapping can't free the physical pages associated that create the
mapping.

This patch makes the higher level page infrastructure not freed when the
kernel stack is temporarily unmapped. This is the final fix needed to
make the kernel consistently boot using the new memory management
changes.