kernel: improve crash dumps
Firstly, even if they are no symbols, we now print out the frame address
anyway. Also, we load symbols using the mmap functionality, which
greatly improves the speed of dumps. Since the vm_object for the
executable is already mapped (as the .text segment), most of the
necessary pages are already cached.

Secondly, display a crash message when the process triggers an
unrecoverable page fault.