kernel: avoid race condition when accessing task state
Both fork() and create_task() needs to access the state of the currently
running task, in order to copy these values into the task they create.
The issue is that these functions accessed the task_state field of the
task, instead of the user_task_state field.

This caused newly created tasks to sometimes seg fault on startup, due
to kernel cs and ss selectors being copied from the task_state, instead
of the userspace versions. This only happened sometimes since the
task_state from the process is only overridden if the system call
happens to get preempted by the scheduler.