kernel: fix support for stopping and continuing tasks
After several changes to blocking and scheduling, this mechanism stopped
working. A new sched_state is added: STOPPED, which replaces the old
WAITING status which is now reserved for blocking and wait queues.

In addition, on system call exit, an additional check is now performed
to see if the task's sched_state has been modified. This will prevent
returning from a system call after a terminating signal is sent, since
now the kernel will notice this state change and run a different task.

Hopefully, this mechanism can be used to prevent resource leaks when
terminating a stopped program, and also to implement execve correctly in
a multithreaded process.