kernel: fix rare double fault when context switching
load_task_into_memory is responsible to setting up all the necessary
pointers needed to be able to call __run_task() to actually do the
context switch. Importantly, it sets up the kernel stack for that
particular task. The double fault was caused by this stack not being
mapped in at all.

As an optimization to this function, load_task_into_memory tried not to
do anything, if the task was already loaded into memory. The issue is
that it used the current_task pointer to detect this condition, but the
current_task is not always loaded into memory. In particular, this
occurs if there are two tasks, 1 currently running and the other trying
to exit. When prompted to reschedule, the scheduler notices one task is
trying to exit and so will switch to that task in order to exit it. Once
finished, the running task will be scheduled again.
load_task_into_memory thought that it didn't need to do work because the
current_task didn't change, but in fact the address space was clobbered
by freeing the other task.

The 'fixed' version always loads things like the kernel stack, but
avoids call load_cr3 if there is no need to do so, since not flushing
the cache was the whole point of the optimization anyway.